Update namespaces, extract/move files
