+++ src/Ooze.Typed.Query/Filters/IOozeQueryFilterProvider.cs
@@ -1,0 +1,6 @@
+﻿namespace Ooze.Typed.Query.Filters;
+
+public interface IOozeQueryFilterProvider<TEntity>
+{
+    IEnumerable<IQueryFilterDefinition<TEntity>> GetFilters();
+}
