+++ src/Ooze.Typed.Query/Filters/IQueryFilterDefinition.cs
@@ -1,0 +1,5 @@
+﻿namespace Ooze.Typed.Query.Filters;
+
+public interface IQueryFilterDefinition<TEntity>
+{
+}
