+++ src/Ooze.Typed.Query/Filters/QueryFilterDefinition.cs
@@ -1,0 +1,9 @@
+﻿using System.Linq.Expressions;
+
+namespace Ooze.Typed.Query.Filters;
+
+internal class QueryFilterDefinition<TEntity> : IQueryFilterDefinition<TEntity>
+{
+    public string Name { get; set; }
+    public LambdaExpression DataExpression { get; set; }
+}
