+++ src/Ooze.Typed.Query/OozeQueryHandler.cs
@@ -1,5 +1,7 @@
 ﻿using Ooze.Typed.Queries;
+using Ooze.Typed.Query.Expressions;
 using Ooze.Typed.Query.Filters;
+using Ooze.Typed.Query.Tokenization;
 
 namespace Ooze.Typed.Query;
